+++ slackerz/classes/string.h
