+++ slackerz/default.h
@@ -2,10 +2,11 @@
 #define _DEFAULT_H_
 
 #include <iostream>
-#include "string.h"
-#include "list.h"
+#include "classes/string.h"
+#include "classes/list.h"
 #include <ctime>
-#include "print.h"
+#include "functions/print.h"
+#include "classes/random.h"
 
 bool init_called = false;
 using namespace slackerz;
@@ -66,36 +67,6 @@ namespace slackerz{
     int abs(int num){
         return abs(num);
     }
-class r{
-    public:
-        int randint(int num = 99999){
-            try{
-                if(init_called == false){
-                    throw "You need to call init_slackerz() to use random\n";
-                }
-
-                return rand() % num + 0;   
-            }
-            catch(const char* msg){
-                cerr << msg;
-                exit(1);
-            }
-        }
-        int randint(int min, int max){
-            try{
-                if(init_called == false){
-                    throw "You need to call init_slackerz() to use random\n";
-                }
-
-                return rand()%(max-min + 1) + min;
-            }
-            catch(const char* msg){
-                cerr << msg;
-                exit(1);
-            }
-        }
-};
-r random;
 //to int
     template <typename U>
     int toint(U something);
